Rush Controller

Rush is the controller in a centralized rush network. The controller starts and stops the workers, pushes tasks to the workers and fetches results.

Local Workers

A local worker runs on the same machine as the controller. Local workers are are spawned with the $start_workers() method via the processx` package.

Remote Workers

A remote worker runs on a different machine than the controller. Remote workers are started manually with the $create_worker_script() method. Remote workers can be started on any system as long as the system has access to Redis and all required packages are installed. Only a heartbeat process can kill remote workers. The heartbeat process also monitors the remote workers for crashes.

Stopping Workers

Local and remote workers can be terminated with the $stop_workers(type = "terminate") method. The workers evaluate the currently running task and then terminate. The option type = "kill" stops the workers immediately. Killing a local worker is done with the tools::pskill() function. Remote workers are killed by pushing a kill signal to the heartbeat process. Without a heartbeat process a remote worker cannot be killed (see section heartbeat).

Heartbeat

The heartbeat process periodically signals that a worker is still alive. This is implemented by setting a timeout on the heartbeat key. Furthermore, the heartbeat process can kill the worker.

Data Structure

Rush writes a task and its result and additional meta information into a Redis hash.

key : xs | ys | extra | state

The key of the hash identifies the task in Rush. The fields are written by different methods, e.g. $push_result() writes ys when the result is available. The value of a field is a serialized list e.g. unserializing xs gives list(x1 = 1, x2 = 2). This data structure allows to quickly convert a hash into a row and to join multiple hashes into a table. For example, three hashes from the above example are converted to the following table.

| key | x1 | x2 | y | timestamp | state   |
| 1.. |  3 |  4 | 7 |  12:04:11 | finished |
| 2.. |  1 |  4 | 5 |  12:04:12 | finished |
| 3.. |  1 |  1 | 2 |  12:04:13 | finished |

Notice that a value of a field can store multiple columns of the table.

The methods $push_tasks() and $push_results() write into multiple hashes. For example, $push_tasks(xss = list(list(x1 = 1, x2 = 2), list(x1 = 2, x2 = 2)) writes xs in two hashes.

Task States

A task can go through four states "queued", "running", "finished" or "failed". Internally, the keys of the tasks are pushed through Redis lists and sets to keep track of their state. Queued tasks are waiting to be evaluated. A worker pops a task from the queue and changes the state to "running" while evaluating the task. When the task is finished, the state is changed to "finished" and the result is written to the data base. If the task fails, the state is changed to "failed"instead of"finished"`.

Queues

Rush uses a shared queue and a queue for each worker. The shared queue is used to push tasks to the workers. The first worker that pops a task from the shared queue evaluates the task. The worker queues are used to push tasks to specific workers.

Fetch Tasks and Results

The $fetch_*() methods retrieve data from the Redis database. A matching method is defined for each task state e.g. $fetch_running_tasks() and $fetch_finished_tasks(). If only the result of the function evaluation is needed, $fetch_results() and $fetch_latest_results() are faster. The methods $fetch_results() and $fetch_finished_tasks() cache the already queried data. The $block_*() variants wait until a new result is available.

Error Handling

When evaluating tasks in a distributed system, many things can go wrong. Simple R errors in the worker loop are caught and written to the archive. The task is marked as "failed". If the connection to a worker is lost, it looks like a task is "running" forever. The methods $detect_lost_workers() and $detect_lost_tasks() detect lost workers. Running these methods periodically adds a small overhead.

Logging

The worker logs all messages written with the lgr package to the data base. The lgr_thresholds argument defines the logging level for each logger e.g. c(rush = "debug"). Saving log messages adds a small overhead but is useful for debugging. By default, no log messages are stored.

Public fields

network_id

(character(1))
Identifier of the rush network.

config

(redux::redis_config)
Redis configuration options.

connector

(redux::redis_api)
Returns a connection to Redis.

processes

(processx::process)
List of processes started with $start_workers().

Active bindings

n_workers

(integer(1))
Number of workers.

n_running_workers

(integer(1))
Number of running workers.

worker_ids

(character())
Ids of workers.

running_worker_ids

(character())
Ids of running workers.

terminated_worker_ids

(character())
Ids of terminated workers.

killed_worker_ids

(character())
Ids of killed workers.

lost_worker_ids

(character())
Ids of lost workers.

tasks

(character())
Keys of all tasks.

queued_tasks

(character())
Keys of queued tasks.

running_tasks

(character())
Keys of running tasks.

finished_tasks

(character())
Keys of finished tasks.

failed_tasks

(character())
Keys of failed tasks.

n_queued_tasks

(integer(1))
Number of queued tasks.

n_queued_priority_tasks

(integer(1))
Number of queued priority tasks.

n_running_tasks

(integer(1))
Number of running tasks.

n_finished_tasks

(integer(1))
Number of finished tasks.

n_failed_tasks

(integer(1))
Number of failed tasks.

n_tasks

(integer(1))
Number of all tasks.

data

(data.table::data.table)
Contains all performed function calls.

worker_info

(data.table::data.table())
Contains information about the workers.

worker_states

(data.table::data.table())
Contains the states of the workers.

priority_info

(data.table::data.table)
Contains the number of tasks in the priority queues.

snapshot_schedule

(character())
Set a snapshot schedule to periodically save the data base on disk. For example, c(60, 1000) saves the data base every 60 seconds if there are at least 1000 changes. Overwrites the redis configuration file. Set to NULL to disable snapshots. For more details see redis.io.

Methods

Public methods

Rush$new()

Rush$format()

Rush$print()

Rush$start_workers()

Rush$restart_workers()

Rush$create_worker_script()

Rush$wait_for_workers()

Rush$stop_workers()

Rush$detect_lost_workers()

Rush$reset()

Rush$read_log()

Rush$push_tasks()

Rush$push_priority_tasks()

Rush$fetch_latest_results()

Rush$wait_for_latest_results()

Rush$fetch_results()

Rush$fetch_queued_tasks()

Rush$fetch_priority_tasks()

Rush$fetch_running_tasks()

Rush$fetch_finished_tasks()

Rush$wait_for_finished_tasks()

Rush$fetch_failed_tasks()

Rush$fetch_tasks()

Rush$wait_for_tasks()

Rush$write_hashes()

Rush$read_hashes()

Rush$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Rush$new(network_id = NULL, config = NULL)

Arguments

network_id

(character(1))
Identifier of the rush network. Controller and workers must have the same instance id. Keys in Redis are prefixed with the instance id.

config

(redux::redis_config)
Redis configuration options. If NULL, configuration set by rush_plan() is used. If rush_plan() has not been called, the REDIS_URL environment variable is parsed. If REDIS_URL is not set, a default configuration is used. See redux::redis_config for details.

Method format()

Helper for print outputs.

Usage

Rush$format(...)

Arguments

...

(ignored).

Returns

(character()).

Method print()

Print method.

Usage

Rush$print()

Returns

(character()).

Method start_workers()

Start workers locally with processx. The processx::process are stored in $processes. Alternatively, use $create_worker_script() to create a script for starting workers on remote machines. By default, worker_loop_default() is used as worker loop. This function takes the arguments fun and optionally constants which are passed in ....

Usage

Rush$start_workers(
  n_workers = NULL,
  wait_for_workers = TRUE,
  globals = NULL,
  packages = NULL,
  heartbeat_period = NULL,
  heartbeat_expire = NULL,
  lgr_thresholds = NULL,
  lgr_buffer_size = 0,
  supervise = TRUE,
  worker_loop = worker_loop_default,
  ...
)

Arguments

n_workers

(integer(1))
Number of workers to be started.

wait_for_workers

(logical(1))
Whether to wait until all workers are available.

globals

(character())
Global variables to be loaded to the workers global environment.

packages

(character())
Packages to be loaded by the workers.

heartbeat_period

(integer(1))
Period of the heartbeat in seconds.

heartbeat_expire

(integer(1))
Time to live of the heartbeat in seconds.

lgr_thresholds

(named character() | named numeric())
Logger threshold on the workers e.g. c(rush = "debug").

lgr_buffer_size

(integer(1))
By default (lgr_buffer_size = 0), the log messages are directly saved in the Redis data store. If lgr_buffer_size > 0, the log messages are buffered and saved in the Redis data store when the buffer is full. This improves the performance of the logging.

supervise

(logical(1))
Whether to kill the workers when the main R process is shut down.

worker_loop

(function)
Loop run on the workers. Defaults to worker_loop_default which is called with fun. Pass fun in ....

...

(any)
Arguments passed to worker_loop.

Method restart_workers()

Restart workers.

Usage

Rush$restart_workers(worker_ids)

Arguments

worker_ids

(character())
Worker ids to be restarted.

Method create_worker_script()

Create script to start workers. The worker is started with start_worker().

Usage

Rush$create_worker_script(
  globals = NULL,
  packages = NULL,
  heartbeat_period = NULL,
  heartbeat_expire = NULL,
  lgr_thresholds = NULL,
  lgr_buffer_size = 0,
  worker_loop = worker_loop_default,
  ...
)

Arguments

globals

(character())
Global variables to be loaded to the workers global environment.

packages

(character())
Packages to be loaded by the workers.

heartbeat_period

(integer(1))
Period of the heartbeat in seconds.

heartbeat_expire

(integer(1))
Time to live of the heartbeat in seconds.

lgr_thresholds

(named character() | named numeric())
Logger threshold on the workers e.g. c(rush = "debug").

lgr_buffer_size

(integer(1))
By default (lgr_buffer_size = 0), the log messages are directly saved in the Redis data store. If lgr_buffer_size > 0, the log messages are buffered and saved in the Redis data store when the buffer is full. This improves the performance of the logging.

worker_loop

(function)
Loop run on the workers. Defaults to worker_loop_default which is called with fun. Pass fun in ....

...

(any)
Arguments passed to worker_loop.

Method wait_for_workers()

Wait until n workers are available.

Usage

Rush$wait_for_workers(n)

Arguments

n

(integer(1))
Number of workers to wait for.

Method stop_workers()

Stop workers.

Usage

Rush$stop_workers(type = "terminate", worker_ids = NULL)

Arguments

type

(character(1))
Type of stopping. Either "terminate" or "kill". If "terminate" the workers evaluate the currently running task and then terminate. If "kill" the workers are stopped immediately.

worker_ids

(character())
Worker ids to be stopped. If NULL all workers are stopped.

Method detect_lost_workers()

Detect lost workers. The state of the worker is changed to "lost". Local workers without a heartbeat are checked by their process id. Checking local workers on unix systems only takes a few microseconds per worker. But checking local workers on windows might be very slow. Workers with a heartbeat process are checked with the heartbeat. Lost tasks are marked as "lost".

Usage

Rush$detect_lost_workers(restart = FALSE)

Arguments

restart

(logical(1))
Whether to restart lost workers.

Method reset()

Stop workers and delete data stored in redis.

Usage

Rush$reset(type = "kill")

Arguments

type

(character(1))
Type of stopping. Either "terminate" or "kill". If "terminate" the workers evaluate the currently running task and then terminate. If "kill" the workers are stopped immediately.

Method read_log()

Read log messages written with the lgr package from a worker.

Usage

Rush$read_log(worker_ids = NULL)

Arguments

worker_ids

(character(1))
Worker ids. If NULL all worker ids are used.

Method push_tasks()

Pushes a task to the queue. Task is added to queued tasks.

Usage

Rush$push_tasks(xss, extra = NULL, terminate_workers = FALSE)

Arguments

xss

(list of named list())
Lists of arguments for the function e.g. list(list(x1, x2), list(x1, x2))).

extra

(list)
List of additional information stored along with the task e.g. list(list(timestamp), list(timestamp))).

terminate_workers

(logical(1))
Whether to stop the workers after evaluating the tasks.

Returns

(character())
Keys of the tasks.

Method push_priority_tasks()

Pushes a task to the queue of a specific worker. Task is added to queued priority tasks. A worker evaluates the tasks in the priority queue before the shared queue. If priority is NA the task is added to the shared queue. If the worker is lost or worker id is not known, the task is added to the shared queue.

Usage

Rush$push_priority_tasks(xss, extra = NULL, priority = NULL)

Arguments

xss

(list of named list())
Lists of arguments for the function e.g. list(list(x1, x2), list(x1, x2))).

extra

(list)
List of additional information stored along with the task e.g. list(list(timestamp), list(timestamp))).

priority

(character())
Worker ids to which the tasks should be pushed.

Returns

(character())
Keys of the tasks.

Method fetch_latest_results()

Fetch latest results from the data base.

Usage

Rush$fetch_latest_results(fields = "ys", data_format = "data.table")

Arguments

fields

(character())
Fields to be read from the hashes.

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Latest results.

Method wait_for_latest_results()

Block process until a new result is available. Returns latest results or NULL if no result is available after timeout seconds.

Usage

Rush$wait_for_latest_results(
  fields = "ys",
  timeout = Inf,
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes.

timeout

(numeric(1))
Time to wait for a result in seconds.

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Latest results.

Method fetch_results()

Fetch results from the data base. Results are cached.

Usage

Rush$fetch_results(
  fields = "ys",
  reset_cache = FALSE,
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to "ys".

reset_cache

(logical(1))
Whether to reset the cache.

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table() Results.

Method fetch_queued_tasks()

Fetch queued tasks from the data base.

Usage

Rush$fetch_queued_tasks(
  fields = c("xs", "xs_extra"),
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra").

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of queued tasks.

Method fetch_priority_tasks()

Fetch queued priority tasks from the data base.

Usage

Rush$fetch_priority_tasks(
  fields = c("xs", "xs_extra"),
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra").

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of queued priority tasks.

Method fetch_running_tasks()

Fetch running tasks from the data base.

Usage

Rush$fetch_running_tasks(
  fields = c("xs", "xs_extra", "worker_extra"),
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra", "worker_extra").

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of running tasks.

Method fetch_finished_tasks()

Fetch finished tasks from the data base. Finished tasks are cached.

Usage

Rush$fetch_finished_tasks(
  fields = c("xs", "ys", "xs_extra", "worker_extra", "ys_extra"),
  reset_cache = FALSE,
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra", "worker_extra", "ys", "ys_extra").

reset_cache

(logical(1))
Whether to reset the cache.

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of finished tasks.

Method wait_for_finished_tasks()

Block process until a new finished task is available. Returns all finished tasks or NULL if no new task is available after timeout seconds.

Usage

Rush$wait_for_finished_tasks(
  fields = c("xs", "ys", "xs_extra", "worker_extra", "ys_extra"),
  timeout = Inf,
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra", "worker_extra", "ys", "ys_extra").

timeout

(numeric(1))
Time to wait for a result in seconds.

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of finished tasks.

Method fetch_failed_tasks()

Fetch failed tasks from the data base.

Usage

Rush$fetch_failed_tasks(
  fields = c("xs", "worker_extra", "condition"),
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra", "worker_extra", "condition".

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of failed tasks.

Method fetch_tasks()

Fetch all tasks from the data base.

Usage

Rush$fetch_tasks(
  fields = c("xs", "ys", "xs_extra", "worker_extra", "ys_extra", "condition", "state"),
  data_format = "data.table"
)

Arguments

fields

(character())
Fields to be read from the hashes. Defaults to c("xs", "xs_extra", "worker_extra", "ys", "ys_extra", "condition", "state").

data_format

(character())
Returned data format. Choose "data.table" or "list". The default is "data.table" but "list" is easier when list columns are present.

Returns

data.table()
Table of all tasks.

Method wait_for_tasks()

Wait until tasks are finished. The function also unblocks when no worker is running or all tasks failed.

Usage

Rush$wait_for_tasks(keys, detect_lost_workers = FALSE)

Arguments

keys

(character())
Keys of the tasks to wait for.

detect_lost_workers

(logical(1))
Whether to detect failed tasks. Comes with an overhead.

Method write_hashes()

Writes a list to redis hashes. The function serializes each element and writes it to a new hash. The name of the argument defines the field into which the serialized element is written. For example, xs = list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4)) writes serialize(list(x1 = 1, x2 = 2)) at field xs into a hash and serialize(list(x1 = 3, x2 = 4)) at field xs into another hash. The function can iterate over multiple lists simultaneously. For example, xs = list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4)), ys = list(list(y = 3), list(y = 7)) creates two hashes with the fields xs and ys. Different lengths are recycled. The stored elements must be lists themselves. The reading functions combine the hashes to a table where the names of the inner lists are the column names. For example, xs = list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4)), ys = list(list(y = 3), list(y = 7)) becomes data.table(x1 = c(1, 3), x2 = c(2, 4), y = c(3, 7)). Vectors in list columns must be wrapped in lists. Otherwise, $read_values() will expand the table by the length of the vectors. For example, xs = list(list(x1 = 1, x2 = 2)), xs_extra = list(list(extra = c("A", "B", "C"))) does not work. Pass xs_extra = list(list(extra = list(c("A", "B", "C"))))` instead.

Usage

Rush$write_hashes(..., .values = list(), keys = NULL, state = NA_character_)

Arguments

...

(named list())
Lists to be written to the hashes. The names of the arguments are used as fields.

.values

(named list())
Lists to be written to the hashes. The names of the list are used as fields.

keys

(character())
Keys of the hashes. If NULL new keys are generated.

state

(character(1))
State of the hashes.

Returns

(character())
Keys of the hashes.

Method read_hashes()

Reads redis hashes written with $write_hashes(). The function reads the values of the fields in the hashes stored at keys. The values of a hash are deserialized and combined into a single list.

Usage

Rush$read_hashes(keys, fields)

Arguments

keys

(character())
Keys of the hashes.

fields

(character())
Fields to be read from the hashes.

Returns

(list of list())
The outer list contains one element for each key. The inner list is the combination of the lists stored at the different fields.

Method clone()

The objects of this class are cloneable with this method.

Usage

Rush$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.