Rush Worker

RushWorker evaluates tasks and writes results to the data base. The worker inherits from Rush.

Note

The worker registers itself in the data base of the rush network.

Super class

rush::Rush -> RushWorker

Public fields

worker_id

(character(1))
Identifier of the worker.

host

(character(1))
Worker is started on a local or remote host.

heartbeat

(callr::r_process)
Background process for the heartbeat.

lgr_buffer

(lgr::AppenderBuffer)
Buffer that saves all log messages.

Active bindings

terminated

(logical(1))
Whether to shutdown the worker. Used in the worker loop to determine whether to continue.

terminated_on_idle

(logical(1))
Whether to shutdown the worker if no tasks are queued. Used in the worker loop to determine whether to continue.

Methods

Public methods

RushWorker$new()

RushWorker$push_running_task()

RushWorker$pop_task()

RushWorker$push_results()

RushWorker$write_log()

RushWorker$set_terminated()

RushWorker$clone()

Inherited methods

rush::Rush$await_tasks()

rush::Rush$await_workers()

rush::Rush$block_finished_tasks()

rush::Rush$block_latest_results()

rush::Rush$create_worker_script()

rush::Rush$detect_lost_tasks()

rush::Rush$detect_lost_workers()

rush::Rush$fetch_failed_tasks()

rush::Rush$fetch_finished_tasks()

rush::Rush$fetch_latest_results()

rush::Rush$fetch_priority_tasks()

rush::Rush$fetch_queued_tasks()

rush::Rush$fetch_results()

rush::Rush$fetch_running_tasks()

rush::Rush$fetch_tasks()

rush::Rush$format()

rush::Rush$print()

rush::Rush$push_priority_tasks()

rush::Rush$push_tasks()

rush::Rush$read_hashes()

rush::Rush$read_log()

rush::Rush$reset()

rush::Rush$start_workers()

rush::Rush$stop_workers()

rush::Rush$write_hashes()

Method new()

Creates a new instance of this R6 class.

Usage

RushWorker$new(
  network_id,
  config = redux::redis_config(),
  host,
  worker_id = NULL,
  heartbeat_period = NULL,
  heartbeat_expire = NULL,
  lgr_thresholds = NULL
)

Arguments

network_id

(character(1))
Identifier of the rush network. Controller and workers must have the same instance id. Keys in Redis are prefixed with the instance id.

config

(redux::redis_config)
Redis configuration options. See redux::redis_config for details.

host

(character(1))
Worker is started on a local or remote host. See Rush for details.

worker_id

(character(1))
Identifier of the worker. Keys in redis specific to the worker are prefixed with the worker id.

heartbeat_period

(integer(1))
Period of the heartbeat in seconds.

heartbeat_expire

(integer(1))
Time to live of the heartbeat in seconds.

lgr_thresholds

(named character() | named numeric())
Logger threshold on the workers e.g. c(rush = "debug").

Method push_running_task()

Push a task to running tasks without queue.

Usage

RushWorker$push_running_task(xss, extra = NULL)

Arguments

xss

(list of named list())
Lists of arguments for the function e.g. list(list(x1, x2), list(x1, x2))).

extra

(list)
List of additional information stored along with the task e.g. list(list(timestamp), list(timestamp))).

Returns

(character())
Keys of the tasks.

Method pop_task()

Pop a task from the queue. Task is moved to the running tasks.

Usage

RushWorker$pop_task(timeout = 1)

Arguments

timeout

(numeric(1))
Time to wait for task in seconds.

Method push_results()

Pushes results to the data base.

Usage

RushWorker$push_results(
  keys,
  yss = list(),
  extra = list(),
  conditions = list(),
  state = "finished"
)

Arguments

keys

(character(1))
Keys of the associated tasks.

yss

(named list())
List of lists of named results.

extra

(named list())
List of lists of additional information stored along with the results.

conditions

(named list())
List of lists of conditions.

state

(character(1))
Status of the tasks. If "finished" the tasks are moved to the finished tasks. If "error" the tasks are moved to the failed tasks.

Method write_log()

Write log message written with the lgr package to the database.

Usage

RushWorker$write_log()

Method set_terminated()

Mark the worker as terminated. Last step in the worker loop before the worker terminates.

Usage

RushWorker$set_terminated()

Method clone()

The objects of this class are cloneable with this method.

Usage

RushWorker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.